Add files via upload
In this report, you can easily see your profit, sales, cost, and many other Key Performance Integrators (KPIs). This Dashboard is used to present the KPIs to the Shareholders of the Business. This is the Role of a business analyst who analyses all data to make the Insights available to the Shareholders. It is also used for market research and analysis.

#MicrosoftExcel #businessAnalyst #Business #Analyst #Analysis
#PowerBI #Dashboardeaningful #Data #DataAnalyst #SQL #ExcelIt